Stream true 16-bit color; reshape wave to pulse-then-rest
Smoothness: effects truncated to 8-bit ints (256 levels), so slow fades
visibly stepped — the bulb holds each discrete level through dim regions
and the bridge barely smooths in entertainment mode. Effects now return
floats and the daemon packs real 16-bit values (65536 levels) per
channel, so fades are smooth.

wave: now a single travelling pulse that fully enters/exits, followed by
a dim rest gap before repeating (period/transit params), instead of a
continuous sinusoid that was always lit somewhere.

+++ hue/effects/candle.py
@@ -71,7 +71,7 @@ def render(t: float, speed: float = 1.0, phase: float = 0.0) -> tuple[int, int, 
     b = 0.03 * brightness
 
     return (
-        int(min(1, max(0, r)) * 255),
-        int(min(1, max(0, g)) * 255),
-        int(min(1, max(0, b)) * 255),
+        min(1.0, max(0.0, r)) * 255.0,
+        min(1.0, max(0.0, g)) * 255.0,
+        min(1.0, max(0.0, b)) * 255.0,
     )
